Refactored the bags_list pallet
