+++ runtime/src/configs/mod.rs
@@ -1,5 +1,6 @@
 pub mod authorship;
 pub mod balances;
+pub mod bags_list;
 pub mod timestamp;
 pub mod transaction_payment;
 pub mod membership;
